upload of many last changes
