+++ configs/create_dataset_multipleSim_mk.cfg
@@ -1,17 +1,14 @@
 [Basics]
-# number of files to shuffle 
-amountoffiles = 3
-LengthFilelist = 5
 #If set to True only frames with neutrinos as primaries are selected
 onlyneutrinoasprimary = True
 # number of MC-paths given, need to match the number of files to be shuffled, needs to start with 0
-MC_path0 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11477/00000-00999/clsim-base-4.0.3.0.99_eff/
-MC_path1 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11981/00000-00999/clsim-base-4.0.3.0.99_eff/
-MC_path2 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11374/00000-00999/clsim-base-4.0.3.0.99_eff/
+MC_path0 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11477/
+MC_path1 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11981/
+MC_path2 = /data/sim/IceCube/2012/filtered/level2/neutrino-generator/11374/
 geometry_file = /data/sim/sim-new/downloads/GCD/GeoCalibDetectorStatus_2012.56063_V0.i3.gz
-out_folder = /data/user/mkronmueller/data/run01
-output_file = neutrinos_of_3_types_001
-dagman_folder = /data/user/mkronmueller/data/run01/dagman
+out_folder = /data/user/mkronmueller/data/run033
+#data_folder = data/user/mkronmueller/data/run033/data
+dagman_folder = /data/user/mkronmueller/data/run033/dagman
 input_shape = auto
 PulseSeriesMap = InIceDSTPulses
 ## give a comma seperated list of foldes in MC_path to be processed, if 'allinmcpath' use all folders
@@ -19,6 +16,8 @@ folder_list = allinmcpath
 ## give a comma seperated list of files to search for in subfolders of MC_path. If 'allinfolder' use all files.
 ### use standart unix syntax for using files matching a special regular expression, e.g. *numu*
 file_list = allinfolder
+# pickle is
+filelist_typ = pickle 
 
 
 ### Define Input Variables Below
